% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compte_NA.R
\name{compte_NA}
\alias{compte_NA}
\title{compte_NA
compte le nombre total de NA et le nombre de NA par variable}
\usage{
compte_NA(df)
}
\arguments{
\item{df}{datagrame}
}
\value{
liste
}
\description{
compte_NA
compte le nombre total de NA et le nombre de NA par variable
}
\examples{
toto <- data.frame(var1 = 1:4, var2 = c("a", "b", NA, "d"), var3 = c(NA, NA, TRUE, FALSE))
compte_NA(toto)
}
